Initial code snapshot copied from Gitlab repo
This is the initial migration of source code from the OpenXR Gitlab
repository to the public Github OpenXR-SDK repo. It is a snapshot of
the source tree (i.e. everything under /src on Gitlab), taken at the
Gitlab tag 'SPEC_0.90.0_FREEZE_PENDING_RELEASE', aka commit SHA
5d1ad822c1bc10a3d3bce1f6186b0afd990fe245.

Build breakage is expected at this commit.